requirements.txt
# Security Note: Monitor requests package for CVEs via https://github.com/psf/requests/security/advisories
pandas>=2.0.0
pyyaml>=6.0.1
requests>=2.28.0
click>=8.0.0
python-dateutil>=2.8.0
python-dotenv>=0.19.0